[Bifrost] Decouples loglet dependencies from bifrost
A restructure the allows finer-grain control over what types loglet providers have access to. This will be exploited more in upcoming PRs.
