+++ crates/bifrost/src/loglet/mod.rs
@@ -8,10 +8,19 @@
 // the Business Source License, use of this software will be governed
 // by the Apache License, Version 2.0.
 
+mod error;
+#[cfg(test)]
+pub mod loglet_tests;
+mod provider;
+pub(crate) mod util;
+
+// exports
+pub use error::*;
+pub use provider::{LogletProvider, LogletProviderFactory};
+
 use std::ops::Add;
 use std::pin::Pin;
 use std::sync::Arc;
-use std::task::Poll;
 
 use async_trait::async_trait;
 use bytes::Bytes;
@@ -19,7 +28,7 @@ use futures::Stream;
 
 use restate_types::logs::{Lsn, SequenceNumber};
 
-use crate::{LogRecord, LsnExt};
+use crate::LogRecord;
 use crate::{Result, TailState};
 
 // Inner loglet offset
@@ -79,6 +88,7 @@ impl SequenceNumber for LogletOffset {
 ///                      ^ Last Committed
 ///                   ^  -- Last released - can be delivered to readers
 ///
+///
 ///  An empty loglet. A log is empty when trim_point.next() == tail.prev()
 ///
 ///       Semantics of offsets
@@ -93,42 +103,6 @@ impl SequenceNumber for LogletOffset {
 ///       0 -> Offset::INVALID
 pub trait Loglet: LogletBase<Offset = LogletOffset> {}
 impl<T> Loglet for T where T: LogletBase<Offset = LogletOffset> {}
-
-/// Wraps loglets with the base LSN of the segment
-#[derive(Clone, Debug)]
-pub struct LogletWrapper {
-    /// The offset of the first record in the segment (if exists).
-    /// A segment on a clean chain is created with Lsn::OLDEST but this doesn't mean that this
-    /// record exists. It only means that we want to offset the loglet offsets by base_lsn -
-    /// Loglet::Offset::OLDEST.
-    pub(crate) base_lsn: Lsn,
-    loglet: Arc<dyn Loglet>,
-}
-
-impl LogletWrapper {
-    pub fn new(base_lsn: Lsn, loglet: Arc<dyn Loglet>) -> Self {
-        Self { base_lsn, loglet }
-    }
-
-    pub async fn create_wrapped_read_stream(
-        self,
-        start_lsn: Lsn,
-    ) -> Result<LogletReadStreamWrapper> {
-        // Translates LSN to loglet offset
-        Ok(LogletReadStreamWrapper::new(
-            self.loglet
-                .create_read_stream(start_lsn.into_offset(self.base_lsn))
-                .await?,
-            self.base_lsn,
-        ))
-    }
-}
-
-impl PartialEq for LogletWrapper {
-    fn eq(&self, other: &Self) -> bool {
-        self.base_lsn == other.base_lsn && Arc::ptr_eq(&self.loglet, &other.loglet)
-    }
-}
 
 #[async_trait]
 pub trait LogletBase: Send + Sync + std::fmt::Debug {
@@ -194,102 +168,3 @@ pub trait LogletReadStream<S: SequenceNumber>: Stream<Item = Result<LogRecord<S,
 }
 
 pub type SendableLogletReadStream<S = Lsn> = Pin<Box<dyn LogletReadStream<S> + Send>>;
-
-#[async_trait]
-impl LogletBase for LogletWrapper {
-    type Offset = Lsn;
-
-    /// This should never be used directly. Instead, use `create_wrapped_read_stream()` instead.
-    async fn create_read_stream(
-        self: Arc<Self>,
-        _after: Self::Offset,
-    ) -> Result<SendableLogletReadStream<Self::Offset>> {
-        unreachable!("create_read_stream on LogletWrapper should never be used directly")
-    }
-
-    async fn append(&self, data: Bytes) -> Result<Lsn> {
-        let offset = self.loglet.append(data).await?;
-        // Return the LSN given the loglet offset.
-        Ok(self.base_lsn.offset_by(offset))
-    }
-
-    async fn append_batch(&self, payloads: &[Bytes]) -> Result<Lsn> {
-        let offset = self.loglet.append_batch(payloads).await?;
-        Ok(self.base_lsn.offset_by(offset))
-    }
-
-    async fn find_tail(&self) -> Result<TailState<Lsn>> {
-        Ok(self
-            .loglet
-            .find_tail()
-            .await?
-            .map(|o| self.base_lsn.offset_by(o)))
-    }
-
-    async fn get_trim_point(&self) -> Result<Option<Lsn>> {
-        let offset = self.loglet.get_trim_point().await?;
-        Ok(offset.map(|o| self.base_lsn.offset_by(o)))
-    }
-
-    // trim_point is inclusive.
-    async fn trim(&self, trim_point: Self::Offset) -> Result<()> {
-        // trimming to INVALID is no-op
-        if trim_point == Self::Offset::INVALID {
-            return Ok(());
-        }
-        let trim_point = trim_point.into_offset(self.base_lsn);
-        self.loglet.trim(trim_point).await
-    }
-
-    async fn read_next_single(&self, from: Lsn) -> Result<LogRecord<Lsn, Bytes>> {
-        // convert LSN to loglet offset
-        let offset = from.into_offset(self.base_lsn);
-        self.loglet
-            .read_next_single(offset)
-            .await
-            .map(|record| record.with_base_lsn(self.base_lsn))
-    }
-
-    async fn read_next_single_opt(
-        &self,
-        from: Self::Offset,
-    ) -> Result<Option<LogRecord<Self::Offset, Bytes>>> {
-        let offset = from.into_offset(self.base_lsn);
-        self.loglet
-            .read_next_single_opt(offset)
-            .await
-            .map(|maybe_record| maybe_record.map(|record| record.with_base_lsn(self.base_lsn)))
-    }
-}
-
-/// Wraps loglet read streams with the base LSN of the segment
-pub struct LogletReadStreamWrapper {
-    pub(crate) base_lsn: Lsn,
-    inner: SendableLogletReadStream<LogletOffset>,
-}
-
-impl LogletReadStreamWrapper {
-    pub fn new(inner: SendableLogletReadStream<LogletOffset>, base_lsn: Lsn) -> Self {
-        Self { inner, base_lsn }
-    }
-}
-
-impl Stream for LogletReadStreamWrapper {
-    type Item = Result<LogRecord<Lsn, Bytes>>;
-
-    fn poll_next(
-        mut self: Pin<&mut Self>,
-        cx: &mut std::task::Context<'_>,
-    ) -> Poll<Option<Self::Item>> {
-        match self.inner.as_mut().poll_next(cx) {
-            Poll::Ready(Some(Ok(record))) => {
-                Poll::Ready(Some(Ok(record.with_base_lsn(self.base_lsn))))
-            }
-            Poll::Ready(Some(Err(e))) => Poll::Ready(Some(Err(e))),
-            Poll::Ready(None) => Poll::Ready(None),
-            Poll::Pending => Poll::Pending,
-        }
-    }
-}
-
-static_assertions::assert_impl_all!(LogletWrapper: Send, Sync, Clone);
